Add Makefile for demo
The Makefile contains commands for building a demo video using the
animation_demo.py script. It assumes that the FLAME, DECA and SAFA model
checkpoint files are present, whose download instructions can be found in
README.md.

This commits adds the necessary static PNG image file.

A note about it is added in README.md